+++ src/extractor.rs
@@ -1,49 +1,46 @@
 use crate::{
     dom,
     error::ReadabilityError,
-    scorer,
-    scorer::{Scorer, DEFAULT_SCORER},
+    scorer::{self, Scorer, ScorerOptions, TopCandidate},
+    utils::{debug_candidate, debug_candidates},
 };
 use html5ever::{parse_document, serialize, tendril::stream::TendrilSink, ParseOpts};
 use log::{debug, trace};
-use markup5ever_rcdom::{RcDom, SerializableHandle};
-#[cfg(feature = "reqwest")]
-use reqwest;
+use markup5ever_rcdom::{Handle, RcDom, SerializableHandle};
 use scorer::Candidate;
-#[cfg(feature = "reqwest")]
-use std::time::Duration;
 use std::{cell::Cell, collections::BTreeMap, default::Default, io::Read, path::Path};
 use url::Url;
 
 #[derive(Debug)]
-pub struct Product {
+pub struct Readable {
     pub title: String,
     pub content: String,
     pub text: String,
 }
 
-/// Fetch website and extract content.
-#[cfg(feature = "reqwest")]
-pub fn scrape(url: &str) -> Result<Product, ReadabilityError> {
-    let client = reqwest::blocking::Client::builder()
-        .timeout(Duration::new(30, 0))
-        .build()?;
-    let mut res = client.get(url).send()?;
-    if res.status().is_success() {
-        let url = Url::parse(url)?;
-        let product = extract(&mut res, &url)?;
-        Ok(product)
-    } else {
-        Err(ReadabilityError::FetchUrl)
-    }
+#[derive(Debug)]
+pub struct Content {
+    pub node: Handle,
+    pub title: String,
 }
 
-/// Extract content and text with a custom [`Scorer`].
-pub fn extract_with_scorer<R>(
+#[derive(Debug, Default)]
+pub struct ExtractOptions<'a> {
+    pub parse_options: ParseOptions,
+    pub scorer_options: ScorerOptions<'a>,
+}
+
+#[derive(Debug, Default)]
+pub struct ParseOptions {
+    pub strict: bool,
+}
+
+/// Extract content from an HTML reader.
+pub fn extract<R>(
     input: &mut R,
     url: &Url,
-    scorer: &Scorer,
-) -> Result<Product, ReadabilityError>
+    opts: ExtractOptions,
+) -> Result<Readable, ReadabilityError>
 where
     R: Read,
 {
@@ -51,74 +48,76 @@ where
         .from_utf8()
         .read_from(input)?;
 
-    if !dom.errors.is_empty() {
+    if opts.parse_options.strict && !dom.errors.is_empty() {
         return Err(ReadabilityError::ParseHtml(dom.errors));
     }
 
-    let mut title = String::new();
-    let mut candidates = BTreeMap::new();
-    let mut nodes = BTreeMap::new();
-    let handle = dom.document.clone();
-    scorer.preprocess(&mut dom, handle.clone(), &mut title);
-    scorer.find_candidates(Path::new("/"), handle.clone(), &mut candidates, &mut nodes);
+    let content = extract_content(&mut dom, url, opts);
 
-    debug!("Found candidates: {}", candidates.values().len());
-    trace!(
-        "Found candidates: {:?}",
-        candidates
-            .values()
-            .map(|candidate| candidate.node.clone())
-            .collect::<Vec<_>>()
-    );
-
-    let mut id: &str = "/";
-    let mut top_candidate: &Candidate = &Candidate {
-        node: handle.clone(),
-        score: Cell::new(0.0),
-    };
-    for (i, c) in candidates.iter() {
-        let score = c.score.get() * (1.0 - scorer::get_link_density(c.node.clone()));
-        c.score.set(score);
-        if score <= top_candidate.score.get() {
-            continue;
-        }
-        id = i;
-        top_candidate = c;
-    }
     let mut bytes = vec![];
-
-    let node = top_candidate.node.clone();
-
-    debug!("Found top candidate: {node:?}");
-
-    scorer.clean(&mut dom, Path::new(id), node.clone(), url, &candidates);
 
     serialize(
         &mut bytes,
-        &SerializableHandle::from(node.clone()),
+        &SerializableHandle::from(content.node.clone()),
         Default::default(),
-    )
-    .ok();
-    let content = String::from_utf8(bytes).unwrap_or_default();
+    )?;
 
     let mut text: String = String::new();
-    dom::extract_text(node.clone(), &mut text, true);
 
-    debug!("Extracted title: {title}");
-    debug!("Extracted content: {content}");
-    debug!("Extracted text: {text}");
+    dom::extract_text(content.node.clone(), &mut text, true);
 
-    Ok(Product {
-        title,
-        content,
+    let content_string = String::from_utf8(bytes).unwrap_or_default();
+
+    debug!("Extracted title: {}", content.title);
+    trace!("Extracted text: {text}");
+    trace!("Extracted content: {content_string}");
+
+    Ok(Readable {
+        title: content.title,
+        content: content_string,
         text,
     })
 }
 
-/// Extract content and text with the default [`Scorer`].
-pub fn extract<R>(input: &mut R, url: &Url) -> Result<Product, ReadabilityError>
-where
-    R: Read,
-{
-    extract_with_scorer(input, url, &DEFAULT_SCORER)
+/// Extract content `Node` from DOM.
+pub fn extract_content(dom: &mut RcDom, url: &Url, opts: ExtractOptions) -> Content {
+    let mut title = String::new();
+    let mut candidates = BTreeMap::new();
+    let mut nodes = BTreeMap::new();
+    let handle = dom.document.clone();
+    let scorer = Scorer::new(opts.scorer_options);
+
+    scorer.preprocess(dom, handle.clone(), &mut title);
+    scorer.find_candidates(Path::new("/"), handle.clone(), &mut candidates, &mut nodes);
+
+    debug!("Found candidates: {}", candidates.values().len());
+    trace!("Found candidates: {:?}", debug_candidates(&candidates));
+
+    let top_candidate = scorer.find_top_candidate(&candidates).unwrap_or_else(|| {
+        TopCandidate::new(
+            "/",
+            Candidate {
+                node: handle.clone(),
+                score: Cell::new(0.0),
+            },
+        )
+    });
+
+    debug!(
+        "Found top candidate: {:?}",
+        debug_candidate(top_candidate.candidate())
+    );
+
+    scorer.clean(
+        dom,
+        Path::new(top_candidate.id()),
+        top_candidate.node().clone(),
+        url,
+        &candidates,
+    );
+
+    Content {
+        node: top_candidate.node().clone(),
+        title,
+    }
 }