+++ src/scorer.rs
@@ -11,10 +11,11 @@ use markup5ever_rcdom::{
     RcDom,
 };
 use regex::Regex;
-use std::{cell::Cell, collections::BTreeMap, path::Path, rc::Rc};
+use std::{borrow::Cow, cell::Cell, collections::BTreeMap, path::Path, rc::Rc};
 use url::Url;
 
 const PUNCTUATIONS_REGEX: &str = r"([、。，．！？]|\.[^A-Za-z0-9]|,[^0-9]|!|\?)";
+// TODO: remove "comment" from unlikely candidates
 const UNLIKELY_CANDIDATES: &str = "combx|comment|community|disqus|extra|foot|header|menu\
      |remark|rss|shoutbox|sidebar|sponsor|ad-break|agegate\
      |pagination|pager|popup|tweet|twitter\
@@ -23,6 +24,7 @@ const LIKELY_CANDIDATES: &str = "and|article|body|column|main|shadow\
                                               |content|hentry";
 const POSITIVE_CANDIDATES: &str = "article|body|content|entry|hentry|main|page\
      |pagination|post|text|blog|story";
+// TODO: remove "comment" and "com" from unlikely candidates
 const NEGATIVE_CANDIDATES: &str = "combx|comment|com|contact|foot|footer|footnote\
      |masthead|media|meta|outbrain|promo|related\
      |scroll|shoutbox|sidebar|sponsor|shopping\
@@ -46,53 +48,98 @@ lazy_static! {
     static ref UNLIKELY: Regex = Regex::new(UNLIKELY_CANDIDATES).unwrap();
     static ref POSITIVE: Regex = Regex::new(POSITIVE_CANDIDATES).unwrap();
     static ref NEGATIVE: Regex = Regex::new(NEGATIVE_CANDIDATES).unwrap();
-    pub static ref DEFAULT_SCORER: Scorer<'static> = Scorer::default();
 }
 
+#[derive(Clone)]
 pub struct Candidate {
     pub node: Rc<Node>,
     pub score: Cell<f32>,
 }
 
-pub struct Scorer<'a> {
+pub struct TopCandidate<'a> {
+    id: Cow<'a, str>,
+    candidate: Cow<'a, Candidate>,
+}
+
+impl<'a> TopCandidate<'a> {
+    pub fn new(id: &'a str, candidate: Candidate) -> Self {
+        Self {
+            id: Cow::Borrowed(id),
+            candidate: Cow::Owned(candidate),
+        }
+    }
+
+    pub fn id(&self) -> &str {
+        self.id.as_ref()
+    }
+
+    pub fn candidate(&self) -> &Candidate {
+        &self.candidate
+    }
+
+    pub fn node(&self) -> &Rc<Node> {
+        &self.candidate.node
+    }
+
+    pub fn score(&self) -> &Cell<f32> {
+        &self.candidate.score
+    }
+}
+
+/// Distribution of the content score among parent nodes.
+#[derive(Debug)]
+pub enum CandidateScore {
+    /// The same weight for all parent nodes.
+    EqualWeight,
+    /// The weight decreases with the level of the parent node.
+    ///
+    /// For example, a parent node will be weighted more than a grandparent.
+    LevelWeight,
+}
+
+#[derive(Debug)]
+pub struct ScorerOptions<'a> {
+    /// The minimum word length of candidates.
+    pub min_candidate_length: usize,
+    /// The maximal number of parent nodes that will be traversed.
+    pub max_candidate_parents: usize,
+    /// Distribution of the content score among parent nodes.
+    pub candidate_score: CandidateScore,
     pub punctuations: &'a Regex,
     pub unlikely_candidates: &'a Regex,
     pub likely_candidates: &'a Regex,
     pub positive_candidates: &'a Regex,
+    pub positive_candidate_weight: f32,
     pub negative_candidates: &'a Regex,
+    pub negative_candidate_weight: f32,
     pub block_child_tags: &'a [&'a str],
 }
 
-impl Default for Scorer<'_> {
+impl Default for ScorerOptions<'_> {
     fn default() -> Self {
         Self {
+            min_candidate_length: 20,
+            max_candidate_parents: 10,
+            candidate_score: CandidateScore::EqualWeight,
             punctuations: &PUNCTUATIONS,
             likely_candidates: &LIKELY,
             unlikely_candidates: &UNLIKELY,
             positive_candidates: &POSITIVE,
+            positive_candidate_weight: 25.0,
             negative_candidates: &NEGATIVE,
+            negative_candidate_weight: 25.0,
             block_child_tags: &BLOCK_CHILD_TAGS,
         }
     }
 }
 
+pub struct Scorer<'a> {
+    options: ScorerOptions<'a>,
+}
+
 impl<'a> Scorer<'a> {
-    pub fn new(
-        punctuations: &'a Regex,
-        likely_candidates: &'a Regex,
-        unlikely_candidates: &'a Regex,
-        positive_candidates: &'a Regex,
-        negative_candidates: &'a Regex,
-        block_child_tags: &'a [&'a str],
-    ) -> Self {
-        Scorer {
-            punctuations,
-            unlikely_candidates,
-            likely_candidates,
-            positive_candidates,
-            negative_candidates,
-            block_child_tags,
-        }
+    pub fn new(options: ScorerOptions<'a>) -> Self {
+        Scorer { options }
     }
 
     pub fn preprocess(&self, dom: &mut RcDom, handle: Handle, title: &mut String) -> bool {
@@ -111,8 +158,8 @@ impl<'a> Scorer<'a> {
             for name in ["id", "class"].iter() {
                 if let Some(val) = dom::attr(name, &attrs.borrow()) {
                     if tag_name != "body"
-                        && self.unlikely_candidates.is_match(&val)
-                        && !self.likely_candidates.is_match(&val)
+                        && self.options.unlikely_candidates.is_match(&val)
+                        && !self.options.likely_candidates.is_match(&val)
                     {
                         return true;
                     }
@@ -162,70 +209,96 @@ impl<'a> Scorer<'a> {
         false
     }
 
+    /// Find candidate tags in DOM node, and distribute score among them.
     pub fn find_candidates(
         &self,
-        id: &Path,
+        node_id: &Path,
         handle: Handle,
         candidates: &mut BTreeMap<String, Candidate>,
         nodes: &mut BTreeMap<String, Rc<Node>>,
     ) {
-        if let Some(id) = id.to_str().map(|id| id.to_string()) {
+        if let Some(id) = node_id
+            .to_str()
+            .map(|candidate_id| candidate_id.to_string())
+        {
             nodes.insert(id, handle.clone());
         }
 
         if self.is_candidate(handle.clone()) {
-            let score = self.calc_content_score(handle.clone());
-            if let Some(c) = id
-                .parent()
-                .and_then(|pid| self.find_or_create_candidate(pid, candidates, nodes))
-            {
-                c.score.set(c.score.get() + score)
-            }
-            if let Some(c) = id
-                .parent()
-                .and_then(|pid| pid.parent())
-                .and_then(|gpid| self.find_or_create_candidate(gpid, candidates, nodes))
-            {
-                c.score.set(c.score.get() + score / 2.0)
-            }
-        }
+            let content_score = self.calculate_content_score(handle.clone());
 
-        if self.is_candidate(handle.clone()) {
-            let score = self.calc_content_score(handle.clone());
-            if let Some(c) = id
-                .to_str()
-                .map(|id| id.to_string())
-                .and_then(|id| candidates.get(&id))
-            {
-                c.score.set(c.score.get() + score)
-            }
-            if let Some(c) = id
-                .parent()
-                .and_then(|pid| pid.to_str())
-                .map(|id| id.to_string())
-                .and_then(|pid| candidates.get(&pid))
-            {
-                c.score.set(c.score.get() + score)
-            }
-            if let Some(c) = id
-                .parent()
-                .and_then(|p| p.parent())
-                .and_then(|pid| pid.to_str())
-                .map(|id| id.to_string())
-                .and_then(|pid| candidates.get(&pid))
-            {
-                c.score.set(c.score.get() + score)
+            let mut current_node_id = Some(node_id.to_path_buf());
+            let mut level = 1;
+
+            // Traverse all parent nodes and distribute content score.
+            while let Some(current_id) = current_node_id {
+                // Break traversal for performance reasons.
+                if level > self.options.max_candidate_parents {
+                    break;
+                }
+
+                if let Some(candidate) =
+                    current_id.to_str().map(|id| id.to_string()).and_then(|id| {
+                        // Only parent nodes are valid candidates.
+                        if current_id != node_id {
+                            self.find_or_create_candidate(Path::new(&id), candidates, nodes)
+                        } else {
+                            None
+                        }
+                    })
+                {
+                    let adjusted_content_score = match self.options.candidate_score {
+                        CandidateScore::EqualWeight => content_score,
+                        CandidateScore::LevelWeight => content_score / (level as f32),
+                    };
+                    candidate
+                        .score
+                        .set(candidate.score.get() + adjusted_content_score);
+
+                    // Ignore candidates above the `body` node.
+                    if dom::get_tag_name(candidate.node.clone()).as_deref() == Some("body") {
+                        break;
+                    }
+                }
+
+                current_node_id = current_id.parent().map(|pid| pid.to_path_buf());
+                level += 1;
             }
         }
 
         for (i, child) in handle.children.borrow().iter().enumerate() {
             self.find_candidates(
-                id.join(i.to_string()).as_path(),
+                node_id.join(i.to_string()).as_path(),
                 child.clone(),
                 candidates,
                 nodes,
             )
         }
+    }
+
+    // TODO: find top candidates with similar score.
+    pub fn find_top_candidate(
+        &self,
+        candidates: &'a BTreeMap<String, Candidate>,
+    ) -> Option<TopCandidate<'a>> {
+        let mut top_candidate: Option<TopCandidate> = None;
+
+        for (id, candidate) in candidates.iter() {
+            let score = candidate.score.get() * (1.0 - get_link_density(candidate.node.clone()));
+            candidate.score.set(score);
+
+            if top_candidate
+                .as_ref()
+                .map_or(true, |top| score > top.candidate.score.get())
+            {
+                top_candidate = Some(TopCandidate {
+                    id: Cow::Borrowed(id),
+                    candidate: Cow::Borrowed(candidate),
+                });
+            }
+        }
+
+        top_candidate
     }
 
     pub fn clean(
@@ -285,11 +358,11 @@ impl<'a> Scorer<'a> {
         useless
     }
 
-    fn calc_content_score(&self, handle: Handle) -> f32 {
+    fn calculate_content_score(&self, handle: Handle) -> f32 {
         let mut score: f32 = 1.0;
         let mut text = String::new();
         dom::extract_text(handle.clone(), &mut text, true);
-        let mat = self.punctuations.find_iter(&text);
+        let mat = self.options.punctuations.find_iter(&text);
         score += mat.count() as f32;
         score += f32::min(f32::floor(text.chars().count() as f32 / 100.0), 3.0);
         score
@@ -303,11 +376,11 @@ impl<'a> Scorer<'a> {
         {
             for name in ["id", "class"].iter() {
                 if let Some(val) = dom::attr(name, &attrs.borrow()) {
-                    if self.positive_candidates.is_match(&val) {
-                        weight += 25.0
+                    if self.options.positive_candidates.is_match(&val) {
+                        weight += self.options.positive_candidate_weight
                     };
-                    if self.negative_candidates.is_match(&val) {
-                        weight -= 25.0
+                    if self.options.negative_candidates.is_match(&val) {
+                        weight -= self.options.negative_candidate_weight
                     }
                 }
             }
@@ -410,14 +483,14 @@ impl<'a> Scorer<'a> {
 
     fn is_candidate(&self, handle: Handle) -> bool {
         let text_len = dom::text_len(handle.clone());
-        if text_len < 20 {
+        if text_len < self.options.min_candidate_length {
             return false;
         }
         let n: &str = &dom::get_tag_name(handle.clone()).unwrap_or_default();
         match n {
             "p" => true,
             "div" | "article" | "center" | "section" => {
-                !dom::has_nodes(handle.clone(), self.block_child_tags)
+                !dom::has_nodes(handle.clone(), self.options.block_child_tags)
             }
             _ => false,
         }
@@ -464,4 +537,90 @@ pub fn get_link_density(handle: Handle) -> f32 {
         link_length += dom::text_len(link.clone()) as f32;
     }
     link_length / text_length
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::utils::{debug_candidates, CandidateTag};
+    use html5ever::{parse_document, tendril::TendrilSink};
+    use std::{fs::File, io::Read};
+
+    #[test]
+    fn test_find_candidates_basic() {
+        let html = r#"
+        <!DOCTYPE html>
+        <html>
+            <head><title>Test Title</title></head>
+            <body>
+                <h1>Welcome</h1>
+                <p>This is a test paragraph with more than 25 characters.</p>
+            </body>
+        </html>"#;
+        let options = ScorerOptions::default();
+        let scorer = Scorer::new(options);
+        let dom = parse_document(RcDom::default(), Default::default())
+            .from_utf8()
+            .read_from(&mut html.as_bytes())
+            .unwrap();
+
+        assert!(dom.errors.is_empty(), "{:?}", dom.errors);
+
+        let mut candidates = BTreeMap::new();
+        let mut nodes = BTreeMap::new();
+
+        scorer.find_candidates(Path::new("/"), dom.document, &mut candidates, &mut nodes);
+
+        let tags = dbg!(debug_candidates(&candidates));
+
+        assert_eq!(candidates.len(), 1);
+        assert!(tags.contains(&CandidateTag::new("body", None, 1.0)));
+    }
+
+    #[test]
+    fn test_find_candidates_comments() {
+        let mut file = File::open("data/comments/input.html").unwrap();
+        let mut html = String::new();
+        file.read_to_string(&mut html).unwrap();
+
+        let options = ScorerOptions {
+            unlikely_candidates: &Regex::new(
+                "combx|community|disqus|extra|foot|header|menu|remark|rss|shoutbox|sidebar|sponsor|ad-break|agegate|pagination|pager|popup|tweet|twitter|ssba",
+            )
+            .unwrap(),
+            negative_candidates: &Regex::new("combx|contact|foot|footer|footnote|masthead|media|meta|outbrain|promo|related|scroll|shoutbox|sidebar|sponsor|shopping|tags|tool|widget|form|textfield|uiScale|hidden").unwrap(),
+            positive_candidates: &Regex::new("article|body|content|entry|hentry|main|page|pagination|post|blog|story").unwrap(),
+            ..Default::default()
+        };
+        let scorer = Scorer::new(options);
+        let dom = parse_document(RcDom::default(), Default::default())
+            .from_utf8()
+            .read_from(&mut html.as_bytes())
+            .unwrap();
+
+        assert!(dom.errors.is_empty(), "{:?}", dom.errors);
+
+        let mut candidates = BTreeMap::new();
+        let mut nodes = BTreeMap::new();
+
+        scorer.find_candidates(Path::new("/"), dom.document, &mut candidates, &mut nodes);
+
+        let tags = dbg!(debug_candidates(&candidates));
+
+        assert_eq!(candidates.len(), 15);
+
+        assert!(tags.contains(&CandidateTag::new("tbody", None, 6.0)));
+        assert!(tags.contains(&CandidateTag::new("tr", Some("tr_2"), 6.0)));
+        assert!(tags.contains(&CandidateTag::new("table", Some("table_2"), 6.0)));
+        assert!(tags.contains(&CandidateTag::new("td", Some("td_0"), 9.0)));
+        assert!(tags.contains(&CandidateTag::new("td", Some("td_1"), 5.0)));
+        assert!(tags.contains(&CandidateTag::new("td", Some("td_2"), 5.0)));
+        assert!(tags.contains(&CandidateTag::new("td", Some("td_3"), 5.0)));
+        assert!(tags.contains(&CandidateTag::new("div", Some("comment_1"), 7.0)));
+        assert!(tags.contains(&CandidateTag::new("div", Some("comment_2"), 7.0)));
+        assert!(tags.contains(&CandidateTag::new("div", Some("comment_3"), 7.0)));
+        assert!(tags.contains(&CandidateTag::new("div", Some("commtext_1"), 7.0)));
+        assert!(tags.contains(&CandidateTag::new("div", Some("commtext_2"), 7.0)));
+        assert!(tags.contains(&CandidateTag::new("div", Some("commtext_3"), 7.0)));
+    }
 }